Find browser's pid by remote_debugging_port.
There is a bug in the original logic: the child process's pid chould be
smaller than the parent. So, we may return the pid of a renderer process
instead of the browser.

This fix change to finding the pid by the remote_debugging_port. It should
be more robust.

BUG=241749
TEST=run on chromeos device

Review URL: https://chromiumcodereview.appspot.com/16322003