Remove TabContentsController from PanelWindowControllerCocoa
We are now getting the native view directly from the web contents and adding it to the content view.

Also fix the shadow issue for inactive panel. This is because we're not longer calling SetAllowOverlappingViews that is in TabContentsController::ensureContentsVisible.

There is one side effect by not calling SetAllowOverlappingViews. When we animate the minimization for the docked panel, we pause at the title-only mode for a short while. At this moment, the web contents view will be visible and overlapp the bottom part of the titlebar due to that the panel is using non-standard taller titlebar. The fix is to detect this and set the view frame manually.

BUG=240803
TEST=existing tests
R=dimich@chromium.org

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=203169

Review URL: https://chromiumcodereview.appspot.com/15177006