Use DownloadItem::IsDone to check for terminal downloads.
When downloads can be resumed, a download transitioning to INTERRUPTED
state is not necessarily entering a terminal state. Therefore, a
download being in IN_PROGRESS state is not a sufficient indicator of
whether a download may eventually complete. Use DownloadItem::IsDone()
to indicate this.

TBR=estade
BUG=7648

Review URL: https://chromiumcodereview.appspot.com/16007017