Make WebRTC bots use webrtc-limited DEPS.
We're switching over to use webrtc-limited instead of
webrtc-internal.
Changing webrtc-internal to be webrtc-limited aligns better with the
planned use of src-limited in Chrome.
This also makes it possible for WebRTC to move internal source code into
webrtc-internal, which is something we plan to do.

BUG=246987
TEST=local build

Review URL: https://chromiumcodereview.appspot.com/16123022