Add 'notify' flag to AudioHandler setters
We need to send notifications when the AudioHandler setters are invoked,
even if the values don't change so that the UI buble is shown. 
To prevent spurrious notifications on startup, pass a 'notify' flag to 
the internal setters instead of checking the stored values.

We still need to track 'muted_' so that ApplyAudioPolicy() does not send
a spurious notification.

BUG=241235

Review URL: https://chromiumcodereview.appspot.com/15785006