[NaCl SDK] Fix archiving of builds on the buildbots.
The recent failure was caused by the shell= parameter
to subprocess.call.  This needs to be different on windows
and non-windows due to the different underlying behavior.

windows: shell=True is needed if you want the executable
to be found in the PATH (this is not true on linux).

linux: shell=True causes commands specified as lists to
fail because the args are passed to the shell and not the
program running in it. (this is not true on windows).

BUG=
R=binji@chromium.org

Review URL: https://codereview.chromium.org/16594002