Handle key sequence for toggling voice search.  This is a no-op until ontogglevoicesearch is added server-side.
This will be used in lieu of the current server-side implementation of this keyboard shortcut to handle cases where Chrome is capturing keystrokes (see internal bug b/9123020).  In the future, we may also use this API to initiate voice search from an icon in the omnibox.

BUG=230417

Review URL: https://chromiumcodereview.appspot.com/15721006