[Android] Make dmprof work with gyp_managed_install
With gyp_managed_install, the shared libraries on the device are in a
subdirectory of /data/local/tmp instead of /data/data or /data/app-lib
(one of those directories will contain symlinks to the actual library).
Update dmprof to treat such a path as though it is an Android path.

Review URL: https://chromiumcodereview.appspot.com/16206024