Use a consistent extension id for unofficial builds.
This will make it easier to test the apps v2 and native messaging
implementations, both of which require a known extension id for the
white-list. Official builds are not affected, since they all end
up with a consistent extension id anyway, via the web-store or the
canary script.

BUG=222855

Review URL: https://chromiumcodereview.appspot.com/16233003