Make IDB better handle origins with really long names.
When an origin causes IDB to attempt to store a directory whose name
doesn't in the filesystem, that's currently treated as a disk corruption.
This patch accounts for it separately and prevents IDB from doing the
usual corruption handling of deleting the LevelDB directory.

BUG=222471

Review URL: https://chromiumcodereview.appspot.com/16537004