Android WebView: write better jarjar rules.
The set of jarjar rules added in r203926 were too confusing; use a
simpler set instead. This remaps everything to be under
com.android.org.chromium whether it was originally from org.chromium or
not. A few of the resulting class names are longer (guava) but many are
shorter, and it's much more comprehensible.

BUG=
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/16044004