Add a path to use CALayers to display content
Add the flag --use-core-animation to enable this path.

Make the RenderWidgetHostViewCocoa layer-backed, and add a
default white layer.

Add a CALayer, softwareLayer_ to draw into when the accelerated
path is disabled. Update the software draw methods to use the CG
calls (this changes the non-CA path as well, but should be equivalent
to previous behavior).

If accelerated compositing is enabled, then add CAOpenGLLayer to
draw accelerated frames (frames using IOSurfaces). If we receive a
software frame, then delete this layer (and the software layer will
appear below it).

This still has a number of issues to resolve before enabling, including
- corrupted bookmark bars
- increased CPU usage (3.5%->5% without compositing, 7%->10% with)
- increased startup time

BUG=245900
NOTRY=True

Review URL: https://chromiumcodereview.appspot.com/16189012