Fix for uninitialized memory valgrind alerts in IndexedDB unit tests
Valgrind doesn't like clang's optimized code in release builds for:

int foo;
if (!InitOrReturnFalse(&foo) || foo < 0) {
  return;
}

The emitted code is checking the value of foo even if the
InitOrReturnFalse() call is returning false, presumably to avoid a
branch. It's a safe optimiation for the compiler to make, but throws
valgrind for a loop.

Initialize foo with a dummy value to suppress the alert.

BUG=247474
R=dgrogan@chromium.org,thestig@chromium.org

Review URL: https://chromiumcodereview.appspot.com/15643015