Fix WebView compositor input handling
Recent changes to compositor input handling deprecated existing hooks for
Android WebView. This patch addresses these changes, allowing the browser
to perform synchronous, in-process input event handling when the UI and
compositor threads are merged. In particular, an InputHandlerManagerClient
allows the embedder to customize delivery of and response to InputEvents.

BUG=241641

Review URL: https://chromiumcodereview.appspot.com/15920002