Move notification timers into a cross-platform location.
Popups stay visible for a short period of time, but that is
not a platform-dependent feature so we can move it out of
views/ code.  Mac will then get the dismiss behavior
automatically.  This does not cause the Mac version to pause
the timers on mouse hover, that will be done in a different
CL.

BUG=238245

Review URL: https://chromiumcodereview.appspot.com/16271009