Only update details for current network
It's not clear to me how this every worked correctly, but it appears to
have been broken for a while at least. I went with the simpler fix of
ignoring updates for networks other than the displayed one because I
don't want to risk missing updates if the JS and C++ get out of sync
somehow. We can improve this when we convert to NetworkStateHandler
(in progress).

BUG=244360

Review URL: https://chromiumcodereview.appspot.com/16676003