Enable initial policies on Android.
This change just flags two policies that are known to work on Android as
supported on that platform. The current version is set to 30 and may change in
the future; enabling these policies for now allows testing and further
developing other parts of the code.

BUG=247641

Review URL: https://chromiumcodereview.appspot.com/16180014