mac notifications: Let <esc> close the settings view and the tray.
To close the settings view, give the back button a key equivalent of \e
(a view controller isn't in the view's responder chain, so implementing
cancelOperation: doesn't help).

To close the tray, override the window's cancelOperation: and order the window
out which will tell the controller that key was resigned, which then informs
the model classes (implementing cancelOperation: on the controller doesn't work
because NSWindow implements cancelOperation: and so won't forward it along the
responder chain).

While here, also remove -forceWindowSizeUpdate which was just another name
for -updateTrayViewAndWindow.

BUG=none
TEST=open tray, open settings. hit esc, settings close. hit esc, tray closes.
R=rsesek@chromium.org

Review URL: https://codereview.chromium.org/16721002