cc: Disable anti-aliasing in software compositor for tiled layers.
When an edge of a quad is internal to a layer, it should not be anti-aliased.
Since we can't control this behaviour right now, we should just disable
anti-aliasing on external edges as well in this case, to avoid rendering
artifacts inside the layer along tile boundaries.

R=enne@chromium.org
BUG=248222

Review URL: https://codereview.chromium.org/16422003