Ensure that the AURA/ASH way of display the Windows 8 on screen keyboard does not interfere with non AURA chrome.
In non AURA chrome on Windows 8 the OSK is displayed by the OS when the accessibility interfaces on fields return information
on whether it is editable etc. In AURA/ASH chrome we display the OSK programmatically. This is done in the ChromeViewHostMsg_FocusedNodeTouched
IPC handler in chrome\browser. This code also executes for non AURA chrome which causes the OSK to popup and vanish.

Fix is to display the OSK only if USE_AURA is defined.

BUG=234934
TBR=cpu@chromium.org

Review URL: https://codereview.chromium.org/16347007