Made cr-dialog-frame not focused by Tab key.
In the situation of pressing the Tab key repeatedly, when cr-dialog-frame has
focus, it looks like focus is disappeared.  This CL change the tabIndex assigned
to cr-dialog-frame from 0 to -1. Elements that have negative tabIndex can be
focused but are not traversed by Tab key.

BUG=239306
TEST=manually

Review URL: https://chromiumcodereview.appspot.com/16123002