Prevent default handling of Enter key when closing prompt dialog.
On confirming a prompt dialog via an "Enter" key press, the dialog may
close and re-focus the element that was active before the dialog opened.
The browser would then take the default action for the event, i.e.
taking the default action on the newly-focused element (such as clicking
the button, opening a combobox menu etc.). This is inappropriate, so
cancel the event to prevent default processing.

BUG=chromium:247646
TEST=Manual: Focus a select element, cause a popup dialog to appear, hit enter. The select dropdown shouldn't open.

Review URL: https://chromiumcodereview.appspot.com/16049018