Don't use __timezone when using uClibc
It appears that plinux.h is a generated file, created by autoconf on a
glibc system, so it naturally outputs a config for glibc.  This patch
changes one of the settings so that it doesn't break when used with
uClibc.

Patch by Mostyn Bramley-Moore

BUG=NONE
TEST=Opera builds work fine.
R=jshin@chromium.org

Review URL: https://codereview.chromium.org/15973007