[sync] Show link to Google dashboard when user chooses to sync nothing
When a user stops sync by choosing to sync nothing via advanced
settings, we should be more clear that this does not delete data from
the server.

This patch changes the status string displayed in the settings page in
the "Sync nothing" state to include a link to the dashboard, as we do
when sync in enabled.

BUG=247039
TEST=Sync nothing and make sure the settings page links to the dashboard.
R=zea@chromium.org

Review URL: https://codereview.chromium.org/16453009