Wait for media file load before checking for GPU process in test.
The navigation completes before the video is loaded, so we have to use a javascript file to wait for a load event to be sure that the video is loaded and should have launched the GPU process. It also has to use a non-empty video, as an empty video never loads.

BUG=237208

Review URL: https://chromiumcodereview.appspot.com/16373015