Add notification when the host successfully starts.
This adds a listener for when first heartbeat is successful, which I think is the best signal that everything is working and the host is able to receive connections.

On Linux, that can send a signal if requested by the caller. On the script side, when running as a daemon, we create a pipe that allows the parent process to wait for that signal (or a permanent failure), and dump all Python log input to the STDERR. This gives users a much better feedback about whether starting the host worked or not.

BUG=179419

Review URL: https://chromiumcodereview.appspot.com/16448005