Make notifications API send bitmaps instead of data URLs to the browser.
Currently we encode the image data downloaded from the net as a data
URL, which still needs to be sent back to a renderer to be converted
into a bitmap.  This patch sends over an ImageData object which is
ready to be put into an SkBitmap (with a RGBA->ARGB conversion).

TBR=sky@chromium.org - Changes for ash/ tests with notifications
BUG=238823,232288,168943

Review URL: https://chromiumcodereview.appspot.com/15927043