Use MMAP_LIST to classify mmap'ed regions instead of STACKTRACES.
We had used only backtraces (written in the STACKTRACES section in dumps)
to classify mmap'ed regions.  The mmap information in STACKTRACES was
mixed with malloc information.  We recently have more detailed mmap
information (related to /proc/pid/maps) in the MMAP_LIST section from
http://crrev.com/202575, http://crrev.com/201255, http://crrev.com/198878
and so on. (The information is now actually duplicated.)

We're now switching to MMAP_LIST for mmap information.  The duplicated
information will be removed after a while.

BUG=244174
NOTRY=True

Review URL: https://chromiumcodereview.appspot.com/16652011