Changed ActivityLog IsEnabled check
The Activity Log should only be enabled if 1) flag is set, or 2) whitelisted extension is installed. Previously, I was  using ExtensionService::IsExtensionEnabled to determine number two. However that will return true if the extension hasn't been installed; it only returns false if the extension is disabled and installed.

Replaced that check with a more reliable way of telling if the whitelisted extension is installed.

BUG=248645

Review URL: https://chromiumcodereview.appspot.com/16816004