cc: Fix and re-enable DrawRenderPassQuadShaderPermutations test
The GLRendererShaderTest.DrawRenderPassQuadShaderPermutations test was
not working correctly since it included solid color quads which would
change the active program. They are not needed since there is not actual
pixel output to look at, instead we should just add the
RenderPassDrawQuad meant to trigger a behaviour and then test for that
behaviour.

Also some of the methods to test for use of a specific program with a
mask instead tested for the program without a mask by mistake.

R=enne
BUG=241318
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/16799004