Omnibox: Turn on Frecency Scoring in HQP By Default (try 2)
This is a copy of
https://codereview.chromium.org/16132011/
which was reverted
( see revert: https://codereview.chromium.org/16860004/ )

This changelist is identical except it corrects an out-of-bounds subscript in ScoredHistoryMatchTest.Scoring.
The error was on line 150 and 151 of chrome/browser/history/scored_history_match_unittest.cc.


Omnibox: Turn on Frecency Scoring in HQP By Default

Also, disable the field trial that was used to help test the
new scoring mode.

Revises tests so they apply to the new scoring code.

I left in the remnants of the HUP-like-scoring code
because I may re-enable that at some point.

BUG=175751
TBR=pkasting,sky

Review URL: https://chromiumcodereview.appspot.com/16887002