gpu: don't MakeCurrent on GetStateFast
When a client is waiting, it's going to spam GetStateFast to the GPU process,
one per other IPC on the channel. This may cause a lot of context switches, and
slowing down progress.
But it turns out we don't need to MakeCurrent on GetStateFast since we're not
sending GL commands. 

BUG=None

Review URL: https://chromiumcodereview.appspot.com/17006002