Ensure the ChunkDemuxer seek callback is invoked when reaching the end of stream.
Seek never completes if no A/V data is provided for the corresponding seek time.
When seeking past the end of stream for either the audio or video stream (or both),
data cannot be provided for that seek time.
EndOfStream should resume the pending seek if needed.

BUG=249356

Review URL: https://chromiumcodereview.appspot.com/16648006