aura: Add code to allow sliding layers in a window using scroll events.
WindowSlider intercepts events on a window, and consumes the scroll events to
slide the layers in a window (as if the layers were stacked on top of
each other in the window). This patch simply adds the class and some tests for
it. A subsequent patch will use this class for implementing overscroll during
a page load.

BUG=196607
TBR=sky@chromium.org

Review URL: https://codereview.chromium.org/16841003