Checked in sources which are normally autogenerated during Mesa's build process.
Generating these sources requires the Python bindings for libxml2 to be installed, and that dependency causes significant problems for Chromium's build process. It is not feasible to add a dependency on a Python module requiring native libraries.

Mesa's XML processors require validation against the DTD, and none of Python's built-in XML processors support validation. The only currently maintained validating XML parsers for Python require native libraries. Therefore it is not possible to rewrite Mesa's XML processors in pure Python.

These sources were generated manually on Linux using a new gyp target. They compile and run successfully on Windows and Mac OS.

This patch is not appropriate to upstream to the Mesa tree.

BUG=238755
TBR=marcheu@chromium.org

Review URL: https://codereview.chromium.org/17028009