Don't use a sleep in the browser's main thread to throttle swapbuffers.
Throttle by delaying sending the ack of each frame to the renderer until
a vsync interval has passed since the last ack was sent.

Explicitly disable throttling when threaded compositing is enabled (it
was supposed to not kick in because the thread proxy will not send new
frames more than once per vsync).

Also move informing the renderer of the vsync properties to be
immediate (not delayed with the ack) and only send this data when
a new frame has been sent (not every time all acks are cleared).
This should not have any material impact, and is to make the flow
easier to follow.

BUG=245900

Review URL: https://chromiumcodereview.appspot.com/16845005