Make the mapping from client id -> secret asynchronous.
In the original design, I intended the pairing registry delegate to maintain
an in-memory cache of client id -> secret, and for this to be invalidated by
a watcher when the on-disk version changes (for example, if the web-app is
used to revoke a pairing). I think an asynchronous model makes more sense,
and eliminates the need for the watcher.

In addition, this CL fixes a bug in the negotiating authenticator when dealing
with authenticators that process messages asychronously.

BUG=156182

Review URL: https://chromiumcodereview.appspot.com/16893002