Return cancellation closure from RequestSender::StartRequestWithRetry.
This is the first step to make each request cancellable from outside.
In coming patches, the cancellation closure will be returned from each 
method of DriveServiceInterface and DriveUploader and as a final goal
it will be used from JobScheduler.

There's a little bit of ugliness in the patch because currently
request cancellation is strongly tied inside RequestRegistry.
All the hacks will go away once this migration completes and
we just remove RequestRegistry.

BUG=231209

Review URL: https://chromiumcodereview.appspot.com/16816018