Search results for OSX App Launcher
This adds a controller for the search results displayed when a user
types in the app list search box. Results display in an NSTableView with
a single column. Each row has an icon on the left, and one or two lines
of formatted text describing the result.

A custom Cell class is used to render the table columns, which generates
an NSAttributedString to draw the rendered text. Selection, borders and
other drawing is coded to match Win/CrOS implementations.

BUG=138633
TEST=Added AppsSearchResultsControllerTest.*. Also manually: inspected
drawing and launching from a search result, end-to-end.

Review URL: https://chromiumcodereview.appspot.com/16472003