|WebContentsViewWin::GetTopLevelNativeWindow()| is perhaps not correct for 3th dev
In content layer, everywhere get the top level window from |WebContentsViewWin::GetTopLevelNativeWindow() |.
In another way,  content has the dependence of toplevel windows, but in the impl of ||WebContentsViewWin::GetTopLevelNativeWindow() |
we just use the hard code |getparent|.  we thinks the toplevel window must be the contents parents, it is not correct for 3th dev.
however content is a Transparent layer for 3th dev. It is not a good designer. content should not suppose toplevel window is content view's parent.

Review URL: https://chromiumcodereview.appspot.com/16323010