sync: Reduce work done to process bookmark changes
The ChangeReorderBuffer and BookmarkChangeProcessor collaborate to
handle bookmark-specific ordering and hierarchy requirements.  With a
bit of help from the directory and some changes to these classes, we can
achieve the same effect as before with less work.

Prior to this CL, the BookmarkChangeProcessor would apply a
position-affecting from sync to the bookmark model by updating all
siblings of the position change in left to right order.  The idea was to
have all the predecessors of the modified item in sync before applying
the change.  The ChangeReorderBuffer helps out by manufacturing fake
changes for siblings of position changes, and delivering them all to the
BookmarkChangeProcessor in syncable::Directory left-to-right order.

This CL works solves the issue by splitting the work up into two
separate passes.  On the first pass, any modified nodes are moved to the
far right within their parent folder.  The second pass iterates over
these modified items in syncable::Directory left-to-right position
order, and moves them to the proper index in the bookmark model.  This
has the same effect as the earlier algorithm: all predecessors are
synced at the time of the final move operation.  This should be much
cheaper than modifying all the siblings of a position change.

This new algorithm also allows us to remove lots of sibling ordering
requirements from the ChangeReorderBuffer and related functions, since
the BookmarkChangeProcessor no longer requires its changes to be
delivered in sibling order.  It also no longer needs to manufacture fake
changes for siblings of position changes.

This CL also includes a few cosmetic changes to the
BookmarkChangeProcessor.

BUG=123429

Review URL: https://chromiumcodereview.appspot.com/16507010