Make sure skia doesn't try to scale images with less than 32 bits per pixel.
The previous attempt to commit this was reverted because ui_unittests stopped working on windows. Since I don't have a windows box, I can't fix the test, so in this CL I am disabling the tests and notifying the people who can fix it.

This log shows what happened when the test was not disabled:
http://build.chromium.org/p/chromium.win/builders/XP%20Tests%20%28dbg%29%281%29/builds/36957/steps/ui_unittests/logs/stdio

BUG=247081

Review URL: https://chromiumcodereview.appspot.com/17100002