Estimate draw duration in SchedulerClient
This adds a DrawDurationEstimate method to SchedulerClient. The
implementation of this method (in ThreadProxy) produces an
estimate based on recent draw times.

This also adds UMA stats measuring the duration of each draw and
the amount by which this duration is underestimated or
overestimated.

BUG=243459

Review URL: https://chromiumcodereview.appspot.com/16574002