Make the shelf stay revealed upon the first edge gesture in immersive fullscreen. This cleans up the code a bit wrt to gestures in immersive fullscreen.
This CL is the less ugly solution to an ugly problem
In immersive fullscreen we want the following shelf auto hide behavior:
- Set the shelf to auto hide for the duration of immersive fullscreen.
- Restore the old auto hide behavior when exiting immersive fullscreen. (If the user changed the auto hide behavior via an edge gesture that sticks)

Restoring the shelf auto hide behavior is super ugly, so we are left with not modifying the auto hide behavior when entering immersive fullscreen. We stick the visibility state to SHELF_AUTO_HIDE to get the auto hide behavior with the mouse.

For gestures, we set gesture_drag_status_ to GESTURE_DRAG_COMPLETE_IN_PROGRESS to set the auto hide state.

The gesture would not make the shelf stick on the first gesture after entering immersive fullscreen because the initial auto_hide_behavior_ != SHELF_AUTO_HIDE_BEHAVIOR_NEVER and
SetAutoHideBehavior(SHELF_AUTO_HIDE_BEHAVIOR_NEVER) on line 469 in the old code would fail because the autohide behavior would not affect the visibility state in immersive fullscreen.

BUG=248680
TEST=ShelfLayoutManagerTest.GestureDrag

Review URL: https://chromiumcodereview.appspot.com/17076005