Files.app: Changed the order of scripts to be loaded from browser_tests.
When loading the background page of test extension that is used from
FileManagerBrowserTest, it may be happened that a test case is started before
test_cases.js is loaded.

This CL swapped the order of test_cases.js and background.js to ensure the test
cases are defined before the test cases are started.

BUT=250711
TEST=none
R=mtomasz@chromium.org

Review URL: https://codereview.chromium.org/16634003