overscroll: Reset the scroll-state when scroll gesture ends.
The trackpad scroll ends with a zero-velocity fling. If the page was scrolled because
of the scroll events, then it would've have started overscroll. But the overscroll-controller
does maintain a state that the page was scrolled. At the end of the scroll, the controller
needs to reset that state, so that subsequent scrolls are processed for overscroll.
However, zero-velocity flings are filtered out by the gesture-event filter. So notify
the overscroll-controller about filtered out gesture events, so that it can reset
its state correctly.

BUG=249808
R=sky@chromium.org

Review URL: https://codereview.chromium.org/17196002