Set max disk cache size correctly.
Currently we set the max shader cache size too late. The disk cache will
be created with a cache size of 0, and the max_cache_size will have no
effect. This CL changes the default from 0 to the desired
gpu::kDefaultMaxProgramCacheMemoryBytes so we always create the disk cache
with the correct size to match the memory cache (currently set to 6MiB).

BUG=248511

Review URL: https://chromiumcodereview.appspot.com/17283002