Support png wallpapers
We use a weaker (from security perspective) image decode here
in order to support png wallpapers. The weaker image decoder will
only used after user logged in. So if it somehow exploited, the
damage stays in one user's profile. (Actually, we use image decoder
that webkit use. So it is as safe/dangeous as user open the png
wallpaper in a browser window)

After decoded the png wallpaper, we will reencode it to jpeg. And 
thereafter, a robust jpeg decoder is used to decode the saved
jpegs to guarantee we can safely display wallpaper pre login stage. 


BUG=237416
TEST=
1. have a png image ready
2. open wallpaper picker, navigate to custom tab
3. change wallpaper to the png file
expected: the png file should set as wallpaper

4. logout
5. verify the wallpaper persist

Review URL: https://chromiumcodereview.appspot.com/16634016