[sync] Reland r206775: Fix mechanism by which chromeos::UserManagerImpl observes ProfileSyncService
This CL relands r206775 after fixing browser_tests failures. Original review and LGTM: https://codereview.chromium.org/17022005

When a user signs in to Chrome OS, and NOTIFICATION_PROFILE_ADDED is
fired, chromeos::UserManagerImpl::Observe() does not register itself as
an observer of ProfileSyncService because the profile is not yet fully
constructed, and therefore, the logged in profile is not identified as
the default profile. This is due to a regression that was introduced in
r176083.

This is bad because UserManagerImpl fails to observe notifications from
ProfileSyncService that are sent out as a result of auth errors. In
fact, this was the root cause of bug http://crbug.com/234164.

This patch fixes the mechanism by which UserManagerImpl registers itself
as an observer of ProfileSyncService, by responding to
NOTIFICATION_LOGIN_USER_PROFILE_PREPARED, which is fired after the
Profile is fully constructed, and the logged in profile is now the
default profile, resulting in UserManagerImpl correctly registering
itself as an observer of ProfileSyncService.

This patch also undoes the earlier fix landed with r206203.

This is the true fix for http://crbug.com/234164.

TBR=xiyuan@chromium.org
BUG=234164, 249918
TEST=Sign in, change password, sign out and back in; Gaia-reauth must be forced.

Review URL: https://chromiumcodereview.appspot.com/16950014