Localize Chrome Apps directory name, and add Chrome Canary and Chromium.
This adds a ".localized" folder and a <locale>.strings file so that OSX
will display the localized folder name.
This also uses a different apps folder for Canary and Chromium.

BUG=168080
TEST=Change the preferred language to something other than English.
Restart Chrome and create a shortcut.
The Chrome Apps folder name should be localized.

Review URL: https://chromiumcodereview.appspot.com/16855005