Two fixes to reduce exceptions on the live instance.
Do not "del data" since it raises an exception when the file is closed. In
python's GC we trust.

Trap DeadlineExceededError when cleaning up an unneeded CloudStorage upload and
leak the objects instead in case of timeout. There isn't much to do in that case
and we don't want the user to upload yet again.

R=iannucci@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/17072004