Add basic infrastructure to generate statistics.
Make all the log entries relating to statistics in a determinate format so it
can be read back and processed automatically. The log format is intentionally
simple so it can be processed without using a regexp yet human readable.

This is much more efficient than touching the datastore to do statistics
bookeeping.

R=iannucci@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/17326002