Calculate colors for avatar label from frame color.
For dark frame colors, we derive a lighter color from the
frame color which is used as a label background, and we chose
black as text color. For light frame colors, we derive a
darker color from the frame color which is used as a label
background, and we chose white as text color.
For screenshots, please take a look at the bug description.

BUG=241377

Review URL: https://chromiumcodereview.appspot.com/16739006