Restrict access to Pepper MediaStream APIs to whitelisted apps.
Adds a AllowPepperMediaStreamAPI method to content::ContentRendererClient to control access to private media stream apis.
Adds an implementation to ChromeContentRendererClient which checks the URL against a whitelist containing a single extension host. This will be set to the Hangouts app's host name when it's available.

BUG=230980

Review URL: https://chromiumcodereview.appspot.com/16455010