Hover effects in views should not be invoked when mouse events are disabled
Make an early return in Widget::OnMouseEvent() if mouse
events are disabled, since in this case the mouse cursor
is not visible and so no new hover effects should be
invoked.

Added the new helper function IsMouseEventsEnabled() to
native_widget_*. This returns true for all native widget
types except for native_widget_aura, where the cursor
client is queried for the mouse event enabled state.

Moved IsMouseHovered() to view.h. This function now returns
false if either a) the mouse cursor is not within the
view's bounds or, b) mouse events are disabled.

BUG=239363

Review URL: https://chromiumcodereview.appspot.com/15738024