Inline win\WebScreenFactoryInfo into RenderWidgetHostViewWin
This initializes the WebKit::WebScreenInfo directly in content. We could use
some of the ui/gfx helpers instead of raw win32 APIs for some of this, but
it doesn't appear we have any code for gathering the display depth or per-component
information or any other consumers of this information.

BUG=237267

Review URL: https://chromiumcodereview.appspot.com/17108014