linux_aura: Make chrome windows drag sources on X11.
This implements the source side of the Xdnd protocol. It also includes some small changes to which object owns the drag selection so that we can do that part entirely asynchronously without spinning up a nested message loop.

This patch only handles text drags and doesn't change the cursor or show a drag image under the cursor. Both of these require some refactoring and this patch is big enough as is.

BUG=130806

Review URL: https://chromiumcodereview.appspot.com/16271006