Add activateContents() to WebContentsDelegateAndroid.
This patch adds a missing Java-side method of the WebContentsDelegate
interface that allows to activate the given WebContents (i.e. select the
tab that corresponds to the WebContents).

This will allow telemetry's tab.Activate() to work on Android once a
corresponding Chrome on Android change is put in place.

BUG=251291

Review URL: https://chromiumcodereview.appspot.com/17415002