Re-use the OSX App Launcher shim for all profiles.
Currently, switching your app list profile will cause a new shim to be
created on the next restart. However, the app list shortcut is the same
for all profiles. This CL avoids prepending the profile path on to the
app bundle name when no profile name is set on the ShortcutInfo.

BUG=138633
TEST=Added WebAppShortcutCreatorTest.CreateAppListShortcut

Review URL: https://chromiumcodereview.appspot.com/17202004