Fix the inactivity monitor in Files.app's gallery.
The previous implementation was unpredictible. Sometimes the inactivity monitor was detecting that the mouse is over tools, sometimes not. The problem was happening when a user was leaving the window by the top or the bottom edge. The inactivity monitor was detecting it as staying of the cursor over the top or the bottom toolbar. This patch resolves this issue by listening for mouseover and mouseout events to detect if a user really is hovering over the tools, including the problematic edge case when leaving the window by the top or the bottom edge.

TEST=Launch gallery. Leave the window by the top or the bottom edge. Wait 3 seconds. The toolbars should disappear.
BUG=251707
R=hirono@chromium.org

Review URL: https://codereview.chromium.org/17171003