Init NaClChromeMainArgs number_of_cores member so apps can size threadpools
The outer sandbox on Linux and OSX was preventing
sysconf(_SC_NPROCESSORS_ONLN) from succeeding, so that NaCl
applications that need the number of processors/cores to, e.g., figure
out the appropriate size for threadpools were getting a bogus value
(-1 when using newlib and 1 when using glibc).

This is a reland of r204389
(https://chromiumcodereview.appspot.com/16047003) which was reverted
because of test failures on Win7 and Aura build bots -- even though
the test passed on a desktop Win7 machine. Consultation with bradn and
ncbray still indicate bot issues, but who knows?

TEST= browser_tests --gtest_filter=*Sysconf*
BUG= 176522

Review URL: https://chromiumcodereview.appspot.com/16761002