[Android] Add GetDescription and defer calculating device_utc_offset
First, only get the device's utc offset when it is actually needed. Calculating
this offset takes about 100ms on a typical device. The build scripts assume
that creating an instance of AndroidCommands is not so expensive. Instead,
defer calculating this value until it is actually requested. There are
currently several cases where the build scripts call adb directly instead of
through android_commands just to avoid this extra cost.

Second, add a GetDescription method to get a (mostly) user-readable description
of the device.

TBR=frankf@chromium.org

Review URL: https://chromiumcodereview.appspot.com/17088005