ContentViewCore: update |tab_crashed_| for lazy url loads.
ContentViewCoreImpl maintains an internal flag: |tab_crashed_| that is set to
true when the backing renderer crashes and back to false when a load or a
reload happens in the tab.

Currently the ContentViewCoreImpl fails to account for the url loads that are
lazily triggered by the NavigationController when the corresponding WebContents
is displayed (that happens for instance when tabs are restored after cold
start).

This patch fixes that by checking for pending lazy loads before showing the
WebContents.

BUG=250825

Review URL: https://chromiumcodereview.appspot.com/17214004