[Mac] Fix the problem that a popup could be closed while it is still in animation.
This occurs when the test code creates a popup and close the collection immediately. The fix is to close the popup without animation for this case.

All the popup animations are coordinated by the popup collection. Added DCHECKs in the popup controller to guard the places that might trigger the 2nd animation while the previous animation has not ended.

BUG=249131
TEST=reenable the disabled test
R=dewittj@chromium.org, rsesek@chromium.org

Review URL: https://codereview.chromium.org/17468007