Start and close app shims based on shell windows.
Start an app's shim when the first shell window is opened instead of
when the extension host is created. Close the shim when the last shell
window is closed. This makes the app shim more consistent with users'
expectations of when the app is 'open'.

BUG=168080
TEST=Set --enable-app-shims.
Find an app that does not open a shell window on launch.
'Create shortcut' to ensure the app shim exists.
Launching the app should not start the shim.
Make the app open a shell window, the shim should start.
Close the window, the shim should close immediately.

Review URL: https://chromiumcodereview.appspot.com/16300003