Add the application's locale to SourceProfile as it is required by the Firefox importer and only available on the browser process' UI thread.
Some plumbing is required to get the locale into the SourceProfile (as it is created on the FILE thread); and some more IPC changes are required to make sure it makes it's way into the import utility process.

This is a prerequisite for OOP import on Linux (https://codereview.chromium.org/17379006/)!

BUG=56816

Review URL: https://chromiumcodereview.appspot.com/17318002