Add a switch to retain video playback on lost focus
In Android, default behavior is to pause video playback when the app
goes background. However, in some apps, it is preferable to retain
video playback even when the app lost the focus. Add a switch for this
behavior so that apps can choose whether to use this behavior or not.

BUG=249985
R=qinmin@chromium.org,ycheo@chromium.org,jochen@chromium.org,joi@chromium.org

Review URL: https://chromiumcodereview.appspot.com/16914005