Translate: Preserve the dialect code of Chinese
When a page whose Content-Language was 'zh' (not 'zh-TW' or 'zh-CN') was opened, even though CLD knew the dielect of the page, the Translate infobar didn't appear.  This CL respects the result of CLD when Content-Language and CLD agree that the page is written in Chinese and the Content-Language doesn't know the dialect.

BUG=250735

Review URL: https://chromiumcodereview.appspot.com/17230003