cc: Trigger BeginFrames that were just missed retroactively
If a SwapAck, SetNeedsBeginFrame, or a Swap occurs *just* after
a BeginFrame is received and skipped by the compositor, this
patch tries the BeginFrame retroactively if it's possible
we might still be able to produce a frame in time.

Also fixes the include guards of begin_frame_args.h.

BUG=243384,243500

Review URL: https://chromiumcodereview.appspot.com/17315011