Add new PPB_nacl_private interface for new translation cache and option in PnaclCoordinator to use it.
The PNaCl translation cache is moving to be implemented in the browser process with a disk_cache, but the PnaclCoordinator in the plugin still controls all the steps required for translation, including requesting a temp file from the browser to write the translated nexe, and determining the cache key from the bitcode's HTTP headers.In the near-term, the PnaclCoordinator will continue to coordinate the process, but the browser will remember which fd is associated with each cache key, and will store the translated nexe in the cache after translation. This CL implements the PnaclCoordinator side, using a stub implementation in the renderer which doesn't actually cache anything yet.

R=jvoung@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3372

Review URL: https://chromiumcodereview.appspot.com/17335009