Generate a unique terminal ID and use it to associate the RDP connection with the attached session.
Instead of trying to use client/server endpoints of the underlying TCP connection to associate the RDP connection with the attached session, this CL passes a unique terminal ID along with other connection parameters to the server. The ID is passed as the working directory for the initial application. Since RdpClient does not specify the initial application to run, the working directory is not used by the RDP server. Chromoting daemon on the other hand can retrieve the value from the session's properties and so it can see which session the RDP client is attached to at the moment. This works even when a different session is attached to the RDP connection.

BUG=249843
R=wez@chromium.org

Review URL: https://codereview.chromium.org/17389009