Make the butter message for deleting hidden after few seconds.
Originally, the timeout value is specified as zero and it means the butter
message will not be disappeared, which is a undocumented behavior.

In addition, if the timeout value is specified, the message is hidden after the
sum of timeout value and HIDE_DELAY_TIME_MS_ milliseconds, which is also
unexpectable.

This CL changed the behavior of timeout and describe it as comments.

BUG=248221
TEST=manually

Review URL: https://chromiumcodereview.appspot.com/17395012