Init NaClChromeMainArgs number_of_cores member so apps can size threadpools
The outer sandbox on Linux (and OSX?) was preventing
sysconf(_SC_NPROCESSORS_ONLN) from succeeding, so that NaCl
applications that need the number of processors/cores to, e.g., figure
out the appropriate size for threadpools were getting a bogus value
(-1 when using newlib and 1 when using glibc).

This is a reland of r207350
(https://chromiumcodereview.appspot.com/16761002) which was reverted
because of test failures on Win XP and Aura build bots -- even though
the test passed on an old Win XP laptop. This includes a bugfix
(thanks bradn) in _snwprintf_s, which *shouldn't* have been the
culprit, but may have been.

The interesting (and smelly) bit is that the same change (modulo minor
variations to get more debugging output) has landed and reverted due
to consistent failures on a small set of bots each time: first it was
XP, next Win7 and Win Aura, and the most recent, XP and Aura; however,
on other Windows bots it runs fine.

Update: bradn checked _snwprintf_s behavior, and the conjecture that
it always puts a NUL at the end of the buffer is correct.  Thus, it
seems likely that stack content variations allowed the test to pass on
most of the Windows bots, and to fail on others.  However, I'll leave
the test disabled for Windows for now, and try re-enablement in another
CL.

The test is disabled for Windows for now, with re-enablement after
more investigation.  See
https://code.google.com/p/chromium/issues/detail?id=252166 which
tracks that.

TEST= browser_tests --gtest_filter=*Sysconf*
BUG=176522

Review URL: https://chromiumcodereview.appspot.com/17448020