Add chrome-native:// scheme.
The new scheme will be used to show pages rendered by the native
platform widgets (Android's view system in particular).

Decoupling the need for a renderer from WebContents was deemed to be
too drastic and risky, so we plan on using a new scheme that is
renderered as about:blank in the renderer and does not force a new
renderer to be created on the next navigation (see SetSite in
DidNavigate).  This will allow us to use the same navigation handling/
site instance security already in place without a bunch of special
handling.

BUG=178915

Review URL: https://chromiumcodereview.appspot.com/15675024