Insert text directly from the virtual keyboard.
This allows the insertion of text that does not directly correspond to
key codes.

The new API introduced is a temporary solution to allow us to continue
prototyping the virtual keyboard and understanding how the
implementation will interact with IMEs until proper IME integration can
be achieved.

BUG=237533

Review URL: https://chromiumcodereview.appspot.com/16972006