Track NPObject ownership by the originating plugins' NPP identifier. [4/6] (Chrome)
This CL makes Chrome-side changes necessary for tracking of
NPObject ownership to out-of-process NPAPI plugin instances.

Ownership information is now included whenever an NPObject
is marshalled across IPC, to be passed when instantiating
NPObjectProxy instances for objects.

In the plugin process each NPObject is tracked to its owning
plugin instance, and browser-owned objects are tracked to a dummy NPP identifier.

This depends upon crrev.com/17220002 and is a prerequisite for crrev.com/17208003.

BUG=152006

Review URL: https://chromiumcodereview.appspot.com/17208003