cc: Move task graph construction to RasterWorkerPool classes.
This removes the need to construct a temporary graph by
RasterWorkerPool classes. It reduces the number of tasks
created and eliminates the need for a dependencies vector
as part of the internal::WorkerPoolTask class.

This is also a prerequisite to implementing a "ready to activate"
signal as that requires as task graph structure that is not a tree.

BUG=247677

Review URL: https://chromiumcodereview.appspot.com/17244003