Roll web page replay 513:515.
This picks up a change which reverts the use-closest-match functionality
to find the closest match using ratio() instead of quick_ratio().

This fixes the ability to sign in to google accounts on Android which
fixes the smoothness_measurement.

Since ratio() is more expensive than quick_ratio(), it may also cause
the page cyclers to regress a little on all platforms. This is not a
chrome regression.

BUG=252786
NOTRY=True
TBR=slamm@google.com

Review URL: https://codereview.chromium.org/16943014