Replace early check for PNaCl with an on-demand check.
Only do this if PNaCl is not installed, or if the
current version is not compatible (for the wrong
architecture).

That way, when we decouple PNaCl component registration
from the --enable-pnacl flag, we won't have everybody doing
an early check when chrome starts.

Actual UI is still to be done/decided (so it's not pretty).
Furthermore, the component updater service delays each
step by N seconds, even during a CheckForUpdateSoon(),
so on a fast connection, most of the delay is the step timer.

BUG=107438
R=cpu@chromium.org, sehr@chromium.org

Review URL: https://codereview.chromium.org/17001003