Under some circumstances, certain TLS connections are dropped by certain remote servers when the TLS ClientHello record exceeds 256 bytes.
This patch changes the number of ECC formats advertised in the ClientHello
to exactly match the same formats advertised by the desktop version of
Chromium during TLS negotiation, netting a savings of approximately 50
bytes in the ClientHello record. This effectively eliminates the occurrence
of the issue.

Patch is named with a 'z' to ensure it is applied after the other patches
in the folder when import_from_android.sh is run, since that script processes
patches in alphabetical order.

R=digit@chromium.org,wtc@chromium.org
BUG:chromium:245500
TEST: 
1. With V25, Visit http://campusstatebank.com 
2. Enter a fictitious username and click "Submit"
3. The "processing login" page appears.
4. Nothing happens. In some cases, the logo will fail to show.
5. With the proposed patch applied, visit http://campusstatebank.com 
6. Enter a fictitious username and click "Submit"
7. The "processing login" page appears.
8. The browser is redirected to a page where the password can be entered.

Contributed by mckev@amazon.com

Review URL: https://chromiumcodereview.appspot.com/17425002