Fix overscroll glow when the FIP dialog is active
The FIP dialog retains focus even when the user scrolls the contents of a page.
Normally, loss of RenderWidgetHost focus would disable the overscroll effect,
preventing the glow from ever showing when the FIP dialog is open. This patch
address the issue by re-enabling the overscroll effect when gestures are
forwarded to the RenderWidgetHost.

BUG=251031

Review URL: https://chromiumcodereview.appspot.com/17413016