Consider IME is associated with the top level widget rather than child widget.
An important thing is that [NativeWidget A] owns Win32 input focus even
when [View X] is logically focused by FocusManager. As a result, an Win32
IME may want to interact with the native view of [NativeWidget A] rather
than that of [NativeWidget B]. This is why we need to call
GetTopLevelWidget() here.

Previous implementation treated [NativeWidget B] as associated with
the IME but it was wrong. This CL fixes the this confusion.

BUG=246534

Review URL: https://chromiumcodereview.appspot.com/17101019