Replace PresentationIdleDelayScale with PresentationScreenDimDelayScale
Presentation mode used to lengthen the idle delay. It now lengthens the
screen dim delay instead. This CL adds a policy that controls the
scaling of the screen dim delay and deprecates the policy that used to
control the scaling of the idle delay.

BUG=248940
TEST=Updated browser and unit tests

Review URL: https://chromiumcodereview.appspot.com/17408010