Omnibox: Add Shortcuts as InstantExtended Provider
We might as well now that InstantExtended is using the native omnibox.  
Besides, it has a positive effect on quality.

For now, I left the framework that allows InstantExtended to use a
different set of providers.  I'll leave it up to Jered or Brett to decide
if they're ready to get rid of it.

BUG=252892

Review URL: https://chromiumcodereview.appspot.com/17563007