[Sync] Have SBH tell SyncManager which types to purge
The SBH now specifies as part of the ConfigureSyncer call which types need to
be purged. This is determined by whether this is a control types config (in
which case all disabled types are purged) or whether any types have been
disabled since the last configuration (as determined by the SBR).

BUG=247115

Review URL: https://chromiumcodereview.appspot.com/17552014